pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "casm-bset"
version = "2.0a1"
authors = [
  { name="CASM developers", email="casm-developers@lists.engr.ucsb.edu" },
]
description = "Constants, definitions, and generically useful tools used by CASM"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "LGPL2.1+"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v2 or later (LGPLv2+)",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "libcasm-configuration>=2.0a4",
    "numpy",
    "scipy",
    "sparse"
]

[project.urls]
Homepage = "https://prisms-center.github.io/CASMcode_docs/"
Repository = "https://github.com/prisms-center/CASMcode_bset"

[build_sphinx]
source-dir = "doc"
build-dir  = "doc/build"
all_files  = 1


[tool.ruff.lint]
select = ["NPY201", "E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["libcasm", "casm"]
